refactor(#121): stage missed edits to modified files
Add the working-tree changes that were part of the same refactor but
not staged in the previous commit: annotation_streamer.py cleanup,
variant_table.py ABC fill, training_set_streamed.py import fix,
test import path updates, and generate_enhanced_training_set.py
import fix.
