+++ hvantk/core/streamers/variant_table.py
@@ -1,14 +1,12 @@
 """VariantTableStreamer — ABC for variant table DataModels.
 
 Per-variant rows keyed by (locus, alleles), arbitrary annotations.
-
-Concrete subclasses live in ``skills/<plugin>/streamer.py``. This stub
-is filled with the full abstract surface in a later PR, when the first
-concrete consumer (ClinVarVariantTableStreamer) is built.
+Concrete subclasses live in ``skills/<plugin>/streamers.py``.
 """
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
+from typing import Iterable
 
 import hail as hl
 
@@ -16,21 +14,32 @@ from hvantk.core.models import AnnotationTable
 
 
 class VariantTableStreamer(ABC):
-    """Per-DataModel operation contract for variant tables (stub)."""
+    """Per-DataModel operation contract for variant tables."""
 
     def __init__(self, artifact: AnnotationTable) -> None:
         self._artifact = artifact
 
     @classmethod
     def from_path(cls, path: str) -> "VariantTableStreamer":
-        """Construct a streamer from a persisted artifact path."""
         from hvantk.core.io import load
         return cls(load(path))
 
     @abstractmethod
     def to_hail(self) -> hl.Table:
-        """Return the underlying Hail Table.
+        """Return the underlying Hail Table for downstream native operations."""
 
-        Stub method retained so subclasses must provide a Hail-level handle.
-        Additional abstract methods land in a later PR.
+    @abstractmethod
+    def filter_to_genes(self, genes: Iterable[str]) -> hl.Table:
+        """Filter variants to those affecting any gene in ``genes``.
+
+        'Affecting' is plugin-specific (ClinVar uses info.GENEINFO; dbNSFP a
+        gene-symbol column).
+        """
+
+    @abstractmethod
+    def filter_by_pathogenicity(self, labels: Iterable[str]) -> hl.Table:
+        """Return variants whose pathogenicity label is in ``labels``.
+
+        The label vocabulary is plugin-specific; the caller passes intended
+        labels (e.g. ['Pathogenic', 'Likely_pathogenic']).
         """
